Fix overzealous generation of invalidClick actions
action.invalidClick events were fired too many times because we did not
thoroughly validate all DOM elements. As an example, suppose we have the
following piece of code

<button><i class="fa fa-play"></i>PLAY</button>

Let's suppose that the user has clicked a button and the exact target is
stored in the $target variable. More often than not, the call
$target.is("button:enabled") will return true because the $target is, in
effect, an enabled button.

On the off-chance that a user clicks the portion of the button occupied
by the <i class="fa fa-play"></i> element, our test fails since the
event target ($target) is no longer a button, even though it is
encapsulated by one.

To fix this, we need to consider what happens when a user interacts with
'valid' inner DOM elements, i.e. elements that are children of elements
that generate a valid click. In our example, we can simply do this by
changing the selector to "button:enabled, button:enabled *".

References #48.