Resolve missing tracking parameters
Certain tracking events in the project-specific analytics are fired
before required tracking parameters are properly configured. To
remedy this, a custom "geotagx-analytics-configured" event is fired to
notify us that the analytics has been properly configured and is ready
for tracking.

Additionally, the "geotagx-analytics-disabled" event notifies us in
case analytics have been disabled.

References #48.