Add a browser compatibility check
A lot of the feedback we got from users were complaints about errors
during task submissions. A bit of digging around and I found that this
was caused by 'history.pushState' (required by pybossa.js), a method
that is not implemented in IE9 and below.

The implemented compatibility check warns a user when one or more
required HTML5 features are not supported by their browser, explicitly
listing said features.